refactor(actions): refactor massif actions
